[MODIFY] #14 - 패키징 수정
